This is a bogus change -- there was a serious typo in the last commit message. It should read:
The CPP flag definition was removed from the Makefile. It is NOW the responsibility of the program compiling the IO_Buffer to supply the correct compiler flag.

This should remove all the errors where this library was compiled with one compiler, the calling code was compiled with another compiler, and then weird things happened